+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 (prerelease)
+
+### Changed
+- Schema versioning now supports multiple envrionments more robustly. Drop your `cassie_schema.versions` table and run `cassie schema:init -v <last schema version applied>` to upgrade.
+- `schema.cql` is replaced by keyspace-agnostic `schema.rb` with robust support for multiple environments. Paves the way to multiple-keyspace support as well. `schema.cql` is no longer needed.
+
 ## 1.0.6 (prerelease)
 ### Added
 - `--trace` option to `cassie` commands to show error backtraces